Rename TxStore -> TxTable
